requirements.txt
aiohttp==3.7.3
aiohttp-jinja2==1.4.2
alembic==1.11.3
async-timeout==3.0.1
asyncpg==0.28.0
attrs==23.1.0
chardet==3.0.4
gino==1.0.1
idna==3.4
Jinja2==3.0.3
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.7
PyYAML==6.0.1
SQLAlchemy==1.3.24
typing_extensions==4.7.1
yarl==1.9.2
